# Improve performance when plucking a single attribute on a loaded relation