@dvandersluis 2018-11-23
@gmcgibbon could you elaborate what you mean by "This would change behaviour for plucks of single loaded columns"?