@gmcgibbon 2018-11-23
Sorry, I mean that this would change behaviour for single column plucks when the relation is loaded.

Further, if you do `SomeModel.all.load.pluck(:col)` it would bypass the attributes API whereas `SomeModel.all.load.pluck(:col, :other_col)` would not. This doesn't make sense to me.