@dvandersluis 2018-11-26
I understand your concern about `pluck` working differently depending on context, and would be good with changing other forms too, I just wanted to improve a single case first. Other than potential inconsistency, is there a downside to bypassing the attributes API here (given that the bypass only happens when we know it is an attribute)?

I think it's kind of ingrained in rails devs that pluck should be the "fast" way of getting attributes, so I wanted to make that true for this case. 😅