@rafaelfranca 2018-11-26
Fast inconsistent results is a no-go. We want fast, sure, but we also want consistency. 

Would the value of `SomeModel.all.load.pluck(:col)` de different of `SomeModel.select(:col).map(&:col)` be different if the attribute has a custom type?