@dvandersluis 2018-11-26
I meant inconsistent on the method, not the results. I can't answer your question @rafaelfranca with 100% certainty, but I believe since `read_attribute` calls `_read_attribute`, the result would be the same, but I'm not sure how to test a custom type.