@dvandersluis 2018-11-26
@gmcgibbon the casting will still happen here:

```
[2] pry(main)> u = User.limit(1).load
  User Load (0.8ms)  SELECT  `users`.* FROM `users` LIMIT 1
=> [#<User id: 3, ...>]
[3] pry(main)> u.pluck(:created_at)
=> [Sun, 11 Oct 2015 08:59:57 UTC +00:00]
```