@gmcgibbon 2018-11-26
Ok, so I think we just lose aliases then in that case. I tested with a [custom typed attribute here](https://gist.github.com/gmcgibbon/73843e5928f04fe64fe462e8bd845f5a) and it seemed to work aside from attribute aliases. I'm fine with this given we start using `_read_attribute` for all loaded plucks, and not just plucks of one column. WDYT @rafaelfranca?