@gmcgibbon 2018-11-26
Actually, this causes another consistency problem with unloaded plucks since they are able to query attribute aliases and loaded plucks aren't with this change... 🤔